feat(*): initial a new branch for next generation
Delete all behaviors scripts to create a new branch for
the next generation of NormaConstructor.
To know more about the next generation, see the documentation for that.

BREAKING CHANGE: all scripts will be rewrite in the next generation

If you are a generator developer, you should rewrite all your generators.
Please read the documentation for the next generation.
And you will see that it's not so diffifult to rewrite generators in new way.